Server refactor: replaced calls to static methods

+++ server/src/main/java/it/smartcommunitylab/validationstorage/service/ArtifactMetadataService.java
@@ -9,22 +9,21 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.ObjectUtils;
 
 import it.smartcommunitylab.validationstorage.common.DocumentNotFoundException;
+import it.smartcommunitylab.validationstorage.common.IdMismatchException;
 import it.smartcommunitylab.validationstorage.common.ValidationStorageUtils;
 import it.smartcommunitylab.validationstorage.model.ArtifactMetadata;
 import it.smartcommunitylab.validationstorage.model.dto.ArtifactMetadataDTO;
 import it.smartcommunitylab.validationstorage.repository.ArtifactMetadataRepository;
-import it.smartcommunitylab.validationstorage.repository.ExperimentRepository;
-import it.smartcommunitylab.validationstorage.repository.ProjectRepository;
 
 @Service
 public class ArtifactMetadataService {
     @Autowired
     private ArtifactMetadataRepository documentRepository;
-
+    
     @Autowired
-    private ProjectRepository projectRepository;
+    private ProjectService projectService;
     @Autowired
-    private ExperimentRepository experimentRepository;
+    private ExperimentService experimentService;
 
     /**
      * Given an ID, returns the corresponding document, or null if it can't be found.
@@ -71,7 +70,7 @@ public class ArtifactMetadataService {
         if (ObjectUtils.isEmpty(projectId))
             throw new IllegalArgumentException("Project ID is missing or blank.");
 
-        ValidationStorageUtils.checkProjectExists(projectRepository, projectId);
+        projectService.findDocumentById(projectId);
 
         String experimentId = request.getExperimentId();
         String runId = request.getRunId();
@@ -87,7 +86,7 @@ public class ArtifactMetadataService {
         documentToSave.setAuthor(author);
 
         // Create experiment document automatically.
-        ValidationStorageUtils.createExperiment(experimentRepository, projectId, experimentId, request.getExperimentName(), author);
+        experimentService.createExperimentIfMissing(projectId, experimentId, request.getExperimentName(), author);
 
         return documentRepository.save(documentToSave);
     }
@@ -115,7 +114,8 @@ public class ArtifactMetadataService {
     public ArtifactMetadata findDocumentById(String projectId, String id) {
         ArtifactMetadata document = getDocument(id);
         if (document != null) {
-            ValidationStorageUtils.checkProjectIdMatch(id, document.getProjectId(), projectId);
+            if (!document.getProjectId().equals(projectId))
+                throw new IdMismatchException();
 
             return document;
         }
@@ -131,7 +131,8 @@ public class ArtifactMetadataService {
         if (document == null)
             throw new DocumentNotFoundException("Document with ID " + id + " was not found.");
 
-        ValidationStorageUtils.checkProjectIdMatch(id, document.getProjectId(), projectId);
+        if (!document.getProjectId().equals(projectId))
+            throw new IdMismatchException();
 
         document.setExperimentId(request.getExperimentId());
         document.setExperimentName(request.getExperimentName());
@@ -146,7 +147,8 @@ public class ArtifactMetadataService {
     public void deleteDocumentById(String projectId, String id) {
         ArtifactMetadata document = getDocument(id);
         if (document != null) {
-            ValidationStorageUtils.checkProjectIdMatch(id, document.getProjectId(), projectId);
+            if (!document.getProjectId().equals(projectId))
+                throw new IdMismatchException();
 
             documentRepository.deleteById(id);
             return;
