+++ server/src/main/java/it/smartcommunitylab/validationstorage/service/DataProfileService.java
@@ -10,12 +10,11 @@ import org.springframework.util.ObjectUtils;
 
 import it.smartcommunitylab.validationstorage.common.DocumentAlreadyExistsException;
 import it.smartcommunitylab.validationstorage.common.DocumentNotFoundException;
+import it.smartcommunitylab.validationstorage.common.IdMismatchException;
 import it.smartcommunitylab.validationstorage.common.ValidationStorageUtils;
 import it.smartcommunitylab.validationstorage.model.DataProfile;
 import it.smartcommunitylab.validationstorage.model.dto.DataProfileDTO;
 import it.smartcommunitylab.validationstorage.repository.DataProfileRepository;
-import it.smartcommunitylab.validationstorage.repository.ExperimentRepository;
-import it.smartcommunitylab.validationstorage.repository.ProjectRepository;
 
 @Service
 public class DataProfileService {
@@ -23,9 +22,9 @@ public class DataProfileService {
     private DataProfileRepository documentRepository;
 
     @Autowired
-    private ProjectRepository projectRepository;
+    private ProjectService projectService;
     @Autowired
-    private ExperimentRepository experimentRepository;
+    private ExperimentService experimentService;
 
     /**
      * Given an ID, returns the corresponding document, or null if it can't be found.
@@ -72,7 +71,7 @@ public class DataProfileService {
         if (ObjectUtils.isEmpty(projectId))
             throw new IllegalArgumentException("Project ID is missing or blank.");
 
-        ValidationStorageUtils.checkProjectExists(projectRepository, projectId);
+        projectService.findDocumentById(projectId);
 
         String experimentId = request.getExperimentId();
         String runId = request.getRunId();
@@ -90,7 +89,7 @@ public class DataProfileService {
         documentToSave.setContents(request.getContents());
 
         // Create experiment document automatically.
-        ValidationStorageUtils.createExperiment(experimentRepository, projectId, experimentId, request.getExperimentName(), author);
+        experimentService.createExperimentIfMissing(projectId, experimentId, request.getExperimentName(), author);
 
         return documentRepository.save(documentToSave);
     }
@@ -118,7 +117,8 @@ public class DataProfileService {
     public DataProfile findDocumentById(String projectId, String id) {
         DataProfile document = getDocument(id);
         if (document != null) {
-            ValidationStorageUtils.checkProjectIdMatch(id, document.getProjectId(), projectId);
+            if (!document.getProjectId().equals(projectId))
+                throw new IdMismatchException();
 
             return document;
         }
@@ -134,7 +134,8 @@ public class DataProfileService {
         if (document == null)
             throw new DocumentNotFoundException("Document with ID " + id + " was not found.");
 
-        ValidationStorageUtils.checkProjectIdMatch(id, document.getProjectId(), projectId);
+        if (!document.getProjectId().equals(projectId))
+            throw new IdMismatchException();
 
         String experimentId = request.getExperimentId();
         String runId = request.getRunId();
@@ -151,7 +152,8 @@ public class DataProfileService {
     public void deleteDocumentById(String projectId, String id) {
         DataProfile document = getDocument(id);
         if (document != null) {
-            ValidationStorageUtils.checkProjectIdMatch(id, document.getProjectId(), projectId);
+            if (!document.getProjectId().equals(projectId))
+                throw new IdMismatchException();
 
             documentRepository.deleteById(id);
             return;
