+++ server/src/main/java/it/smartcommunitylab/validationstorage/service/RunEnvironmentService.java
@@ -10,11 +10,10 @@ import org.springframework.util.ObjectUtils;
 
 import it.smartcommunitylab.validationstorage.common.DocumentAlreadyExistsException;
 import it.smartcommunitylab.validationstorage.common.DocumentNotFoundException;
+import it.smartcommunitylab.validationstorage.common.IdMismatchException;
 import it.smartcommunitylab.validationstorage.common.ValidationStorageUtils;
 import it.smartcommunitylab.validationstorage.model.RunEnvironment;
 import it.smartcommunitylab.validationstorage.model.dto.RunEnvironmentDTO;
-import it.smartcommunitylab.validationstorage.repository.ExperimentRepository;
-import it.smartcommunitylab.validationstorage.repository.ProjectRepository;
 import it.smartcommunitylab.validationstorage.repository.RunEnvironmentRepository;
 
 @Service
@@ -23,9 +22,9 @@ public class RunEnvironmentService {
     private RunEnvironmentRepository documentRepository;
 
     @Autowired
-    private ProjectRepository projectRepository;
+    private ProjectService projectService;
     @Autowired
-    private ExperimentRepository experimentRepository;
+    private ExperimentService experimentService;
 
     /**
      * Given an ID, returns the corresponding document, or null if it can't be found.
@@ -71,7 +70,7 @@ public class RunEnvironmentService {
     public RunEnvironment createDocument(String projectId, RunEnvironmentDTO request, String author) {
         if (ObjectUtils.isEmpty(projectId))
             throw new IllegalArgumentException("Project ID is missing or blank.");
-        ValidationStorageUtils.checkProjectExists(projectRepository, projectId);
+        projectService.findDocumentById(projectId);
 
         String experimentId = request.getExperimentId();
         String runId = request.getRunId();
@@ -89,7 +88,7 @@ public class RunEnvironmentService {
         documentToSave.setContents(request.getContents());
 
         // Create experiment document automatically.
-        ValidationStorageUtils.createExperiment(experimentRepository, projectId, experimentId, request.getExperimentName(), author);
+        experimentService.createExperimentIfMissing(projectId, experimentId, request.getExperimentName(), author);
 
         return documentRepository.save(documentToSave);
     }
@@ -117,7 +116,8 @@ public class RunEnvironmentService {
     public RunEnvironment findDocumentById(String projectId, String id) {
         RunEnvironment document = getDocument(id);
         if (document != null) {
-            ValidationStorageUtils.checkProjectIdMatch(id, document.getProjectId(), projectId);
+            if (!document.getProjectId().equals(projectId))
+                throw new IdMismatchException();
 
             return document;
         }
@@ -133,7 +133,8 @@ public class RunEnvironmentService {
         if (document == null)
             throw new DocumentNotFoundException("Document with ID " + id + " was not found.");
 
-        ValidationStorageUtils.checkProjectIdMatch(id, document.getProjectId(), projectId);
+        if (!document.getProjectId().equals(projectId))
+            throw new IdMismatchException();
 
         String experimentId = request.getExperimentId();
         String runId = request.getRunId();
@@ -150,7 +151,8 @@ public class RunEnvironmentService {
     public void deleteDocumentById(String projectId, String id) {
         RunEnvironment document = getDocument(id);
         if (document != null) {
-            ValidationStorageUtils.checkProjectIdMatch(id, document.getProjectId(), projectId);
+            if (!document.getProjectId().equals(projectId))
+                throw new IdMismatchException();
 
             documentRepository.deleteById(id);
             return;
